rename script modules
