+++ src/assets/scripts/common/navigation.js
