+++ src/assets/scripts/common/preload.js
