+++ src/assets/scripts/common/register-serviceworker.js
