+++ src/assets/scripts/common/signup.js
