+++ src/assets/scripts/common/utils.js
